requirements.txt
autopep8==1.5
beautifulsoup4==4.8.2
entrypoints==0.3
flake8==3.7.9
mccabe==0.6.1
Pillow==7.1.1
pycodestyle==2.5.0
pyflakes==2.1.1
PyPDF2==1.26.0
soupsieve==2.0
